fix(sdd): let review act after implementation, not during it
Receipt-driven review acts on the finished result: you explore,
implement, verify, and only then is review offered, freezing what is
already done. This package implements exactly that, twice.
applyReviewOfferRouting fires in one window — verify has passed and the
kill switch is on — and with the switch off it "returns before doing
anything at all [...] which is what makes the absence structural".
resolveNextRecommended never routes to review before verify.

Two rules disagreed with both, and each was a reported dead end.

Finish refused to close a PASSING IMPLEMENTATION attempt whenever the
review binding covered the pre-attempt bytes. It always does: changing
the candidate is what an attempt is for. That is review deciding whether
implementation may finish, before any verification has run, and it is
#1993 — a maintainer-authorized correction passed every gate and could
not be closed, while the only named exit demanded an approved review of
the corrected candidate that the review side was structurally unable to
produce, because minting it runs NewCompactState and opens a fresh
correction budget and a fresh reviewer the contract forbids after the
one independent verification.

Nothing is given up. The delivery gates re-derive their verdict from the
candidate actually being delivered, so an unreviewed candidate is still
refused there, after SDD finishes. The binding and receipt stay recorded
as ledger metadata: review stops deciding, it does not stop being
tracked.

The second rule made the kill switch a half-measure. Unmanaged
remediation required the switch to be off AND no binding to exist, so a
binding created before the operator turned review off kept blocking the
correction path. The contract on ReviewDisabled says that while review
is off it "does not exist, so it must have no implications", and a
leftover binding is an implication. The replay mirror in
applyRuntimeFinishEvent carried the same condition and is aligned with
it, or a legitimately committed record would make the chain unreplayable.

The machinery that served only the removed gate is deleted rather than
left uncalled: runtimeRemediationExitRefusal,
runtimeStrandedSuccessorRefusal, runtimeRemediatesArgument,
runtimeSelfSuccessorAvailable, RuntimeStrandedSuccessor,
runtimeStrandedSuccessor, the ErrRuntimeRemediationSuccessorRequired
sentinel and its compact mapping. Leaving it in place is how the
coupling grows back.

Closes #1993
