+++ .refusal-ratchet-baseline.txt
@@ -1425,7 +1425,6 @@ internal/sddstatus/review_binding.go	"authority or live gate changed before bind
 internal/sddstatus/review_binding.go	"authority or receipt changed before compact binding selection"
 internal/sddstatus/review_binding.go	"bound compact authority cannot produce its receipt"
 internal/sddstatus/review_binding.go	"bound compact post-apply gate context changed"
-internal/sddstatus/review_binding.go	"bound compact predecessor is stale or not approved"
 internal/sddstatus/review_binding.go	"bound compact predecessor receipt changed"
 internal/sddstatus/review_binding.go	"bound compact predecessor receipt does not match authority"
 internal/sddstatus/review_binding.go	"compact post-apply gate changed before binding selection"
@@ -1474,10 +1473,8 @@ internal/sddstatus/runtime_ledger.go	"SDD runtime record predecessor cycle detec
 internal/sddstatus/runtime_ledger.go	"SDD runtime record revision mismatch: expected %s, got %s"
 internal/sddstatus/runtime_ledger.go	"SDD runtime request identifier was reused with different inputs"
 internal/sddstatus/runtime_ledger.go	"SDD runtime reset request digest does not match record"
-internal/sddstatus/runtime_ledger.go	"a bound passing SDD runtime attempt must also bind the approved review of its corrected candidate"
 internal/sddstatus/runtime_ledger.go	"an atomic remediation successor is valid only for a passed attempt"
 internal/sddstatus/runtime_ledger.go	"approved SDD remediation successor changed before native commit"
-internal/sddstatus/runtime_ledger.go	"approved SDD remediation successor does not bind the natively charged candidate"
 internal/sddstatus/runtime_ledger.go	"approved SDD self-remediation requires distinct corrected evidence"
 internal/sddstatus/runtime_ledger.go	"atomic SDD remediation successor requires a populated native binding"
 internal/sddstatus/runtime_ledger.go	"atomic SDD runtime remediation request digest does not match record"
@@ -1490,7 +1487,6 @@ internal/sddstatus/runtime_ledger.go	"begin record does not continue the termina
 internal/sddstatus/runtime_ledger.go	"begin record exceeds the persisted objective budget"
 internal/sddstatus/runtime_ledger.go	"begin record is not a valid successor"
 internal/sddstatus/runtime_ledger.go	"binding record expected revision does not equal replay state"
-internal/sddstatus/runtime_ledger.go	"bound SDD review change does not match the runtime objective"
 internal/sddstatus/runtime_ledger.go	"committed SDD runtime HEAD did not replay to candidate revision"
 internal/sddstatus/runtime_ledger.go	"duplicate runtime request identifier"
 internal/sddstatus/runtime_ledger.go	"existing immutable SDD runtime record differs from its revision"
