+++ bench/journeys_sdd.go
@@ -1177,6 +1177,37 @@ func sddNamesReviewRouter(observation Observation) bool {
 
 // sddBlockedLeafFinish drives the block and holds the leaf branch's contract:
 // the topology really is a leaf, and the refusal named a finish command.
+// sddBoundPassingFinishCloses is j37's assertion after #1993: a passing
+// implementation attempt closes even though the binding covers the bytes from
+// before the correction. Review acts after implementation and verification,
+// and the delivery gates enforce on the finished candidate.
+func sddBoundPassingFinishCloses(r *journeyRun) error {
+	status, err := readRuntimeStatus(r)
+	if err != nil {
+		return err
+	}
+	// Deliberately NOT sddPassingFinish: that helper asserts the block this
+	// journey exists to prove is gone.
+	observation := r.run(sddAttemptArgs(r, "finish", status.Revision, "bench-finish-bound-passing",
+		append([]string{"--outcome", "passed", "--evidence-revision", sddCorrectedEvidence}, sddTerminalEvidence...)...), false)
+	if observation.ExitCode != 0 {
+		return fmt.Errorf("a bound passing finish over a corrected candidate was refused: %s", firstLine(observation.Stderr))
+	}
+	settled, err := proveRuntime(r.sandbox)
+	if err != nil {
+		return err
+	}
+	if settled.ActiveAttempt != nil {
+		return errors.New("the finish ran but the attempt is still active")
+	}
+	// The binding stays recorded: review stops deciding, it does not stop
+	// being tracked, and the delivery gates read it from here.
+	if settled.BindingRevision == "" {
+		return errors.New("the finish dropped the review binding; it must stay recorded on the ledger")
+	}
+	return nil
+}
+
 func sddBlockedLeafFinish(r *journeyRun) error {
 	if err := proveLeafTopology(r.sandbox, r.sandbox.Scratch["bound"]); err != nil {
 		return err
@@ -1638,76 +1669,19 @@ func sddJourneys() []Journey {
 	return []Journey{
 		// ------------------------------------------ remediation successor cycle
 		{
-			ID:     "j37-sdd-remediation-self-successor",
-			Title:  "Bound passing attempt over a corrected candidate: the block, and the exit it names",
-			Source: "shape 4 (a refusal naming something that does not work) + community deadlock report",
-			// Expected: the plain passing finish is REFUSED — closing a bound
-			// attempt as passed while the binding approves older bytes would
-			// launder unreviewed content into a passing runtime record — and the
-			// refusal names the one finish that is accepted from here, which the
-			// next step then runs. The whole point of the journey is that the
-			// named exit runs: a refusal that names a command refused one layer
-			// deeper is the defect this branch exists to avoid.
-			Steps: []Step{
-				{Name: "fixture: repository with a committed OpenSpec change", Fixture: sddRuntimeRepo},
-				{Name: "begin, fail, begin again", Requires: sddAttemptBeginCapability, Composite: sddBeginFailBegin},
-				{Name: "fixture: the bounded correction moves the candidate", Fixture: sddBoundedCorrection},
-				{Name: "review start on the corrected candidate", Requires: startCapability, Args: productArgs("review", "start"), After: rememberLineage},
-				{Name: "review finalize", Requires: finalizeCapability, Args: productArgs("review", "finalize"), After: rememberLineage},
-				{Name: "bind the approved review to the change", Requires: bindSDDCapability, Composite: sddBindApprovedReview},
-				{Name: "plain passing finish over a changed candidate", Requires: sddAttemptFinishCapability, Composite: sddBlockedLeafFinish},
-				{Name: "the self-successor finish the refusal named", Requires: sddAttemptRemediationCapability,
-					Composite: sddRemediationFinish("", "bench-finish-self-successor")},
-				{Name: "prove the objective closed", Composite: sddProveObjectiveComplete},
-			},
-		},
-		{
-			ID:     "j38-sdd-remediation-distinct-successor",
-			Title:  "Same block with a real recovery successor in the way: the refusal routes to review, not to a finish",
-			Source: "shape 3 (the same guard with a different precondition underneath) + shape 4",
-			// Expected: once `review recover` mints a successor, the bound lineage
-			// is no longer the compact recovery leaf and its post-apply gate stops
-			// allowing, so the self-successor finish would be refused one layer
-			// deeper. The refusal must therefore name the review router instead —
-			// and must NOT name a finish. Following the router verbatim approves
-			// the successor, and the distinct-successor finish then completes.
-			Steps: []Step{
-				{Name: "fixture: repository with a committed OpenSpec change", Fixture: sddRuntimeRepo},
-				{Name: "begin, fail, begin again", Requires: sddAttemptBeginCapability, Composite: sddBeginFailBegin},
-				{Name: "fixture: the bounded correction moves the candidate", Fixture: sddBoundedCorrection},
-				{Name: "review start on the corrected candidate", Requires: startCapability, Args: productArgs("review", "start"), After: rememberLineage},
-				{Name: "review finalize", Requires: finalizeCapability, Args: productArgs("review", "finalize"), After: rememberLineage},
-				{Name: "bind the approved review to the change", Requires: bindSDDCapability, Composite: sddBindApprovedReview},
-				{Name: "fixture: the scope widens after the binding", Fixture: sddWidenScope},
-				{Name: "recover the successor the gate names", Requires: recoverCapability, Composite: sddRecoverSuccessor},
-				{Name: "plain passing finish with a successor in the way", Requires: sddAttemptFinishCapability, Composite: sddBlockedNonLeafFinish},
-				{Name: "run the review transition the refusal routed to", Requires: statusCapability, Composite: executeNextTransitionVerbatim},
-				{Name: "the distinct-successor finish", Requires: sddAttemptRemediationCapability,
-					Composite: sddRemediationFinish(sddSuccessorLineage, "bench-finish-distinct-successor")},
-				{Name: "prove the objective closed", Composite: sddProveObjectiveComplete},
-			},
-		},
-		{
-			ID:     "j39-sdd-remediation-stranded-successor",
-			Title:  "The successor can never be finalized: what the router names, and what actually clears it",
-			Source: "shape 4 (the named continuation runs and changes nothing) + shape 2 (a clearable state read as terminal)",
-			// The successor's candidate is reverted below its own frozen target,
-			// so it can never be finalized. This journey was added while that
-			// state routed to the review router, whose printed transition RAN,
-			// exited 0 and changed nothing, after which the finish it asked for
-			// was refused naming nothing at all. It reported the gap for exactly
-			// one measurement before the gap was closed.
+			ID:     "j37-sdd-bound-passing-attempt-closes-over-a-corrected-candidate",
+			Title:  "Bound passing attempt over a corrected candidate closes; review enforces at delivery",
+			Source: "community deadlock report (#1993)",
+			// This journey used to prove the opposite: that the plain passing
+			// finish was REFUSED, and that the refusal named a self-successor
+			// finish which then ran. Review acts after implementation and
+			// verification, so there is no block here to name an exit for. The
+			// binding stays recorded and the delivery gates enforce on the
+			// finished candidate.
 			//
-			// Expected now: the refusal names the abandonment that clears it,
-			// complete except for the reason and actor the operator supplies, and
-			// names neither the router nor a finish, because both are dead ends
-			// on this shape. Following it restores the approved predecessor as
-			// the leaf, and the finish then succeeds through the self-successor
-			// exit.
-			//
-			// The authorization still has to be assembled by hand, so the cost
-			// shows up honestly in manual_tokens rather than being hidden by the
-			// block having a name now.
+			// j38 (the refusal routing to the review router) and j39 (the
+			// stranded-successor exit) are deleted with the refusal that was
+			// their whole subject.
 			Steps: []Step{
 				{Name: "fixture: repository with a committed OpenSpec change", Fixture: sddRuntimeRepo},
 				{Name: "begin, fail, begin again", Requires: sddAttemptBeginCapability, Composite: sddBeginFailBegin},
@@ -1715,15 +1689,7 @@ func sddJourneys() []Journey {
 				{Name: "review start on the corrected candidate", Requires: startCapability, Args: productArgs("review", "start"), After: rememberLineage},
 				{Name: "review finalize", Requires: finalizeCapability, Args: productArgs("review", "finalize"), After: rememberLineage},
 				{Name: "bind the approved review to the change", Requires: bindSDDCapability, Composite: sddBindApprovedReview},
-				{Name: "fixture: the scope widens after the binding", Fixture: sddWidenScope},
-				{Name: "recover the successor the gate names", Requires: recoverCapability, Composite: sddRecoverSuccessor},
-				{Name: "fixture: strand the successor below its own frozen target", Fixture: sddStrandSuccessor},
-				{Name: "plain passing finish with a stranded successor in the way", Requires: sddAttemptFinishCapability, Composite: sddBlockedStrandedFinish},
-				{Name: "the review transition still creates a real lineage here", Requires: statusCapability, Composite: sddTransitionCreatesALineage},
-				{Name: "abandon the stranded successor, as the refusal named", Requires: abandonCapability, Composite: sddAbandonStrandedSuccessor},
-				{Name: "the self-successor finish, now that the strand is gone", Requires: sddAttemptRemediationCapability,
-					Composite: sddRemediationFinish("", "bench-finish-after-abandon")},
-				{Name: "prove the objective closed", Composite: sddProveObjectiveComplete},
+				{Name: "the plain passing finish closes over the changed candidate, and keeps the binding", Requires: sddAttemptFinishCapability, Composite: sddBoundPassingFinishCloses},
 			},
 		},
 		{
